+++ src/components/Hub/HubCard.vue
@@ -164,7 +164,7 @@
 </template>
 
 <script>
-import { EventBus } from "@/event-bus";
+import { goToSearch } from "@/utils/search";
 const arrford = require("arrford");
 import { format, parseISO } from "date-fns";
 import { getImageURL } from "@/services/Image";
@@ -240,12 +240,10 @@ export default {
 
   methods: {
     categoryClick(e) {
-      //console.log("chip click: ", e.target.innerHTML);
-      let opts = {
+      goToSearch(this.$router, {
         query: e.target.innerText.toLowerCase(),
         type: "hub",
-      };
-      EventBus.$emit("search", opts);
+      });
     },
     errorHandler(id) {
       console.log("error for image: ", id);